// ldk/node/dist/loopClients/whisperSensor.d.ts
export interface WhisperDisambiguationElement<T extends string> {
    type: T;
    /**
     *  Value the UI uses to order the form inputs.
     *  Value must be greater than 0
     *  If this value is ommited it will default to 0
     */
    order?: number;
}
export declare type WhisperDisambiguationOption = WhisperDisambiguationElement<'option'> & {
    label: string;
};
export declare type WhisperDisambiguationText = WhisperDisambiguationElement<'text'> & {
    body?: string;
};
export declare type WhisperDisambiguationElements = WhisperDisambiguationOption | WhisperDisambiguationText;
// ...
export declare enum WhisperListStyle {
    NONE = 0,
    SUCCESS = 1,
    WARN = 2,
    ERROR = 3
}
export declare enum WhisperListAlign {
    LEFT = 0,
    CENTER = 1,
    RIGHT = 2
}